+++ include/Math.h
@@ -4,25 +4,25 @@
 
 // --- Matrix operations
 
-std::vector<std::vector<float>> operator*(std::vector<std::vector<float>>& firstMatrix,
-    std::vector<std::vector<float>>& secondMatrix);  // Matrix - Matrix multiplication - overloaded operator
+std::vector<std::vector<double>> operator*(std::vector<std::vector<double>>& firstMatrix,
+    std::vector<std::vector<double>>& secondMatrix);  // Matrix - Matrix multiplication - overloaded operator
 
-std::vector<std::vector<float>> multiplyMatrices(std::vector<std::vector<float>>& firstMatrix,
-    std::vector<std::vector<float>>& secondMatrix); // Matrix - Matrix multiplication
+std::vector<std::vector<double>> multiplyMatrices(std::vector<std::vector<double>>& firstMatrix,
+    std::vector<std::vector<double>>& secondMatrix); // Matrix - Matrix multiplication
 
-std::vector<float> operator*(std::vector<std::vector<float>>& Matrix, std::vector<float>& Vector); // Matrix - Vector multiplication - overloaded operator
+std::vector<double> operator*(std::vector<std::vector<double>>& Matrix, std::vector<double>& Vector); // Matrix - Vector multiplication - overloaded operator
 
-std::vector<float> multiplyMatrixVector(std::vector<std::vector<float>>& Matrix, std::vector<float>& Vector); // Matrix - Vector multiplication
+std::vector<double> multiplyMatrixVector(std::vector<std::vector<double>>& Matrix, std::vector<double>& Vector); // Matrix - Vector multiplication
 
 bool randBool();
  
-float randProb();
+double randProb();
 
-int preSort(std::vector<std::vector<float>>& orderedVector);
+int preSort(std::vector<std::vector<double>>& orderedVector);
 
-void quickSort(std::vector<std::vector<float>>& orderedVector, int startingPoint, int endingPoint);
+void quickSort(std::vector<std::vector<double>>& orderedVector, int startingPoint, int endingPoint);
 
-int partition(std::vector<std::vector<float>>& orderedVector, int startingPoint, int endingPoint); 
+int partition(std::vector<std::vector<double>>& orderedVector, int startingPoint, int endingPoint); 
 
-std::vector<float> mtMm(std::vector<std::vector<float>>& Matrix, std::vector<float>& Vector);
+std::vector<double> mtMm(std::vector<std::vector<double>>& Matrix, std::vector<double>& Vector);
